split the rest of RenderUtils into WorldRenderUtils
